+++ src/Public/New-LocalInstance.ps1
@@ -26,13 +26,21 @@ function New-LocalInstance {
         # Specifies the name of the instance to create.
         [Parameter()]
         [ValidateNotNullOrEmpty()]
-        [string] $Name = "test-$( New-DatabaseName )"
+        [string] $Name = ( New-DatabaseName ),
+
+        # Specifies the version of the sql server.
+        [Parameter()]
+        [System.Version] $Version
     )
 
-    Import-Module PsSqlLocalDb -ErrorAction Stop
+    Import-Module PsSqlLocalDb -MinimumVersion 0.3 -ErrorAction Stop
 
     # create instance
-    $instance = New-LocalDbInstance -Name $Name
+    if ( $Version ) {
+        $instance = New-LocalDbInstance -Name $Name -Version $Version
+    } else {
+        $instance = New-LocalDbInstance -Name $Name
+    }
 
     # add metadata
     $instance | Add-Member 'DataSource' "(LocalDb)\$( $instance.Name )"